Moved unlinkChallengeTasks to challenge model and added tests
